Refactoring; Added volume and frequency charts.
